requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.12.23
Pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
